requirements.txt
aiohttp==3.11.18
APScheduler==3.11.0
nest_asyncio==1.6.0
python-dotenv==1.1.0
python-telegram-bot==22.0
Requests==2.32.3
